+++ src/components/Form/FormWrapper/FormWrapper.styled.tsx
@@ -17,7 +17,7 @@ export const FormWrapperStyled = styled.form`
     &.spacing-medium ${FormItemStyled}:not(:last-child) { margin-bottom : 32px; }
     &.spacing-large  ${FormItemStyled}:not(:last-child) { margin-bottom : 40px; }
     &.spacing-huge   ${FormItemStyled}:not(:last-child) { margin-bottom : 48px; }
-/* 
+
     &.spacing-none   ${FormItemGroupStyled} ${FormItemStyled}:last-child { margin-bottom :    0; }
     &.spacing-nano   ${FormItemGroupStyled} ${FormItemStyled}:last-child { margin-bottom :  8px; }
     &.spacing-micro  ${FormItemGroupStyled} ${FormItemStyled}:last-child { margin-bottom : 12px; }
@@ -25,5 +25,5 @@ export const FormWrapperStyled = styled.form`
     &.spacing-small  ${FormItemGroupStyled} ${FormItemStyled}:last-child { margin-bottom : 24px; }
     &.spacing-medium ${FormItemGroupStyled} ${FormItemStyled}:last-child { margin-bottom : 32px; }
     &.spacing-large  ${FormItemGroupStyled} ${FormItemStyled}:last-child { margin-bottom : 40px; }
-    &.spacing-huge   ${FormItemGroupStyled} ${FormItemStyled}:last-child { margin-bottom : 48px; } */
+    &.spacing-huge   ${FormItemGroupStyled} ${FormItemStyled}:last-child { margin-bottom : 48px; }
 `;